+++ src/Services/Identity/Identity.API/MailTemplates/SendLastLoginInfoEmailTemplate.html
@@ -1,0 +1,19 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Last Login Info</title>
+
+
+</head>
+<body>
+    <p>[Gender] [FullName]</p>
+    <br />
+    <br />
+    <p>Your account was signed in at [LoginDate] from [IpAddress]</p>
+    <br />
+    <p>You are getting this email to make sure it was you</p>
+    <br />
+    <p>If you don't recognize this login, you should change your password immediately</p>
+</body>
+</html>
