+++ src/Services/Identity/Identity.Application/Features/Commands/Users/SignInWithTwoFactorCommand/SignInWithTwoFactorCommand.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text.Json.Serialization;
-using Identity.Application.Features.Commands.ViewModels;
+using Identity.Application.Features.Commands.Users.ViewModels;
 using MediatR;
 using Olcsan.Boilerplate.Utilities.Results;
