+++ src/Services/Identity/Identity.Application/Mappings/MappingProfile.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using Identity.Application.Features.Commands.Users.CreateUserCommand;
 using Identity.Application.Features.Commands.Users.SignUpCommand;
-using Identity.Application.Features.Commands.Users.UpdateAddressFromUserCommand;
+using Identity.Application.Features.Commands.Users.UpdateAddressCommand;
 using Identity.Application.Features.Queries.ViewModels;
 using Identity.Domain.Entities;
