+++ src/Services/Identity/Identity.Domain/Enums/Gender.cs
@@ -1,0 +1,9 @@
+﻿namespace Identity.Domain.Enums
+{
+    public enum Gender
+    {
+        Male,
+        Female,
+        Other,
+    }
+}
