+++ src/Services/Identity/Identity.Domain/Enums/Role.cs
@@ -1,6 +1,6 @@
 ﻿namespace Identity.Domain.Enums
 {
-    public enum Roles
+    public enum Role
     {
         Administrator,
         Editor,
